+++ SlashBar/Modules/Settings/SettingsBook.cs
@@ -1,0 +1,5 @@
+namespace SlashBar.Modules.Settings;
+
+public static class SettingsBook {
+    public static SettingsStore Store { get; } = new();
+}